requirements.txt
asgiref==3.5.2
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
Pillow==9.2.0
PyJWT==1.7.1
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.3.0
